fix: Fix broken pivot table on MultiIndex and simplify code
Fix an error in pivot_table where linkages may get carried over from
the original LDF, depending on pandas implementation details, which was
exposed in the groupby fix.

Change logic to construct a fresh pandas.DataFrame, add the relevant
columns,
and pivot that.

Add regression test which is a modification of the original `test_pivot_table`
using MultiIndexes on the linked tables.