Fix error when aggregating linked bool columns
Between pandas 0.24 and 1.4, aggregating on linked bool columns broke
due to the linkage attempting to apply a NaN mask to the bool datatype.
This  impacted getting counts of linked records satisfying specific
conditions (e.g. `link.sum(attribute == 3)`).

There was existing logic to apply a compatible mask to columns with an
integer datatype, so this simply adds checks for bool and uint datatypes
to that.

Add a regression test for this feature.