fix problem with missing index dimension
`np.nonzero(arr)` (equivalent to `np.where(arr)`) returns an array for each
dimension of the input `arr`. Since `flat_indexer` is 1-D, it returns a
tuple of length 1 with a single array in it. We just want that array.

Also update to use `np.nonzero` as recommended by numpy documentation.