fix missing index update logic
Reverse the order of arrays in `np.isin`, and wrap with `np.nonzero`,
to correctly get the indices of the new dataframe which had missing
values in the original.

This fixes a long-standing logic error, which was not noticed before
due to a separate issue which meant `missing_indices` were never
actually computed correctly.